Fix bug about Wi-Fi dialog rotation
When user clicks a Wi-Fi access point in WifiSettings,
screen pops up a Wi-Fi point dialog. And then user
rotates the screen, Wi-Fi access dialog changes to
"Add network" full screen dialog.

In old code, we check whether dialog is showing by
dialog.isShowing() in onSaveInstanceState.
For now, this design is not appropriate. Since isShowing()
won't return true anymore when fragment calls onSaveInstanceState.

So, we check dialog object whether is null or not now.
If dialog is null, it means that there is no dialog was shown,
before user rotates the screen.

Change-Id: I7dc26369c005f576fe679abc70327f6a02620935
Fixes: 112624846
Test: manual test, robo test

Former-commit-id: b53cdf09ddd04b098b949cec4f6bce661f8587cd