Move content later button doesn't work on Migration Wizard
"Move content later" button doesn't work on StorageWizardMigrateConfirm
screen which is shown when end-user try to migrate data from adopted
SD Card to internal storage.

The cause is "Move content later" button triggers to StorageWizardReady
screen but cannot launch because the mDisk is null when the destination
of migration is internal storage.

This CL fixes to just exit StorageWizardMigrateConfirm screen when mDisk
is null.

Bug: 117082495
Test: manual
Change-Id: Iccdccf4dda126d77458b9db35e4ec6ae6a263cb7

Former-commit-id: 901934372894f8dcc6cd7aacafcd7607a4515845