Update the layout of half width suggestion cards.
When switching to the dismissal view, the background color of suggestion
cards that were built as half cards with only 1 line text in it won't
fully cover the whole card. The alignments are not consistent as well.

Make layout_height of ViewFlipper and the main content to be
match_parent.

Change-Id: If345d00d21150e184614029a34e79b3210ec52ff
Fixes: 122559038
Test: visual

Former-commit-id: 26127c627148d4be88155739a5b7e76a2a2c0d34