Remove the multi-SIM bottom nav in network and internet v2 mode
When a device has multiple SIMs, at the bottom of the SIM details page
we show a chooser with an icon for each SIM card. This UX design is
being supplanted by the new network and internet v2 design where there
is a list of networks you can get to from the Network & internet
homepage. So when the v2 mode is turned on, we don't want that bottom
navigation to appear.

Bug: 116349402
Test: visual (will remove existing tests for this feature eventually)
Change-Id: I9cae479a1f4e689adcc4f4262661b10edbbbf2d3

Former-commit-id: 33340366aa5270358c54538de38bd20f7d2cdf74