Add dialog for changing display name for mobile network
This CL adds a pencil/edit icon at the top of the mobile network details
page. When clicked, it brings up a dialog that allows the user to set
the display name used for that mobile network - this is most useful in
multi-SIM situations. This dialog also displays the mobile network
operator name and phone number which can help the user keep straight
which one they are renaming.

Bug: 122670283
Test: make RunSettingsRoboTests
Change-Id: Ifd63cfb66ce743bd5ba8a06b52229f06ed56b7bd

Former-commit-id: e13d04c762eaa16b6d56adebd6270983f3614fa5