Move trust agent toggles to developer options
This CL moves two toggles for trust agent extend unlock mode from
Security > Settings to Developer Options instead. It also updates some
of the configuration strings to reflect that this toggles behavior for
trust agents in general.

No new logic is being added here, this simply moves options intended
only for internal dogfooding to a more appropriate location in
preparation for the public Beta.

Bug: 111435975
Bug: 120871688
Test: Manually verified that the settings show up in Developer Options
and work as intended.

Change-Id: I2b6705d50fa783089a5c0dfabf76677af44209f7

Former-commit-id: 8f53ad0f3785cd7bfda0ad2bed84351f21f93301