Add extra to the intent for adding additional subscription
In the places where we have a link to let the user add an additional
mobile network subscription, we need the "fore provision" extra added to
the intent.

Bug: 127532036
Test: manual (enable DSDS mode and the settings_network_internet_v2
feature flag, then go to Settings->Network & internet and try clicking
on the '+' button on the right of the Mobile network preference, or
click through that pref if you have 2+ sims and hit the 'Add more'
link. This should bring you into an 'add sim' flow provided by the eSIM
manager)

Change-Id: I56ab2125938f2df07f756d100a7ac993b1e5ced3

Former-commit-id: 68d56353fb2d7b7175575a9cbde23da9bccaafd5