Improve UX of WifiDppChooseSavedWifiNetworkFragment
In WifiDppChooseSavedWifiNetworkFragment, it does not show the saved Wi-Fi
network which is not reachable, users may be confused about why their saved
Wi-Fi network not on the list. This change shows saved networks regardless
of signal reachable or not.

Fix WifiDppChooseSavedWifiNetworkFragment UI overlap problem, remove redundant
ScrollView and adjust ConstraintLayout parameters.

Bug: 124129645
Test: manual test
Change-Id: I86637e8722bb8f40a3b995a617dc1e22b2ae324a

Former-commit-id: d0ceb83c95b2039982fee5a3da2bb12f493d2964