Add Sensors Off QS developer tile
This CL will replace the SensorPrivacyTile with a new tile that can be
enabled from the developer settings. When this new tile is enabled only
the camera, mic, and sensors controlled by the SensorManager will be
disabled; the location and airplane mode tiles will not be modified.
The user will be notified when this tile is enabled with the sensors
off icon in the status bar.

Fixes: 126618519
Test: Manually verified the Sensors Off tile was available in the QS page
      after being enabled from the developer settings.

Change-Id: I028aa1c836b00e6a8d129e46a925d2d2c2d61759

Former-commit-id: 3a665ac5ecf601f371ba8b17e309be554b3e67d4